Only poll for readiness in the 'prepare' step
There's no need to pass a 'IOCondition' to the helper: we only need to
know when it's possible to send an operation without blocking. Further
processing should be wrapped in a 'IOSchedulerJob' or a 'ThreadPool'.

Setup the file descriptors in 'prepare' since instances might be added
after the creation of the source.

Remove locks, it's up to the caller to not reset servers while they are
traversed.